refactor: bỏ field Points, chấm điểm theo thang 10 và bỏ trừ điểm khi chấm

+++ src/Services/Exam/Exam.Contracts/ExamResultAdminListItemDto.cs
@@ -6,7 +6,7 @@ public record ExamResultAdminListItemDto(
     string Email,
     string FullName,
     decimal TotalScore,
-    int MaxPossibleScore,
+    decimal MaxPossibleScore,
     bool? Passed,
     DateTime ExamStartDate,
     DateTime? ExamFinishDate,
